third_party/WebKit/LayoutTests/imported/web-platform-tests/html/semantics/embedded-content/media-elements/synchronising-multiple-media-elements/media-controllers/task-source-expected.txt

This is a testharness.js-based test.
FAIL MediaController events task source assert_array_equals: property 1, expected object "[object MediaController]" but got object "[object MediaController]"
Harness: the test ran to completion.
